+++ sdk/spring/spring-cloud-azure-core/src/main/java/com/azure/spring/core/resource/AzureStorageResource.java
@@ -3,16 +3,16 @@
 
 package com.azure.spring.core.resource;
 
+import java.net.URLConnection;
+import java.util.Locale;
 import org.springframework.core.io.AbstractResource;
 import org.springframework.core.io.WritableResource;
 import org.springframework.lang.NonNull;
 import org.springframework.util.Assert;
-
-import java.util.Locale;
+import org.springframework.util.StringUtils;
 
 /**
- * Abstract implementation of {@link WritableResource} for reading and writing objects in Azure
- * StorageAccount.
+ * Abstract implementation of {@link WritableResource} for reading and writing objects in Azure StorageAccount.
  *
  * @author Warren Zhu
  */
@@ -28,6 +28,14 @@ abstract class AzureStorageResource extends AbstractResource implements Writable
         assertIsAzureStorageLocation(location);
         int containerEndIndex = assertContainerValid(location);
         return location.substring(getProtocolPrefix().length(), containerEndIndex);
+    }
+
+    String getContentType(String location) {
+        String objectName = getFilename(location);
+        if (StringUtils.hasText(objectName)) {
+            return URLConnection.guessContentTypeFromName(objectName);
+        }
+        return null;
     }
 
     String getFilename(String location) {